Migrate Mutation.deleteAlertChannels
